create new files for constants and utils
